+++ changelog.md
@@ -10,6 +10,7 @@
 
 - Support Llama 3.1 8B instruct
 - Use models directly from Huggingface with config.use_hf_model()
+- Add "echo" config option to allow streaming tokens to stdout as they are generated
 
 ## 0.21 - 2024-09-25